#7 refactor: split controllers into restControllers and pageControllers packages
