+++ src/main/java/eventify/pageControllers/AdminBookingPageController.java
@@ -1,4 +1,4 @@
-package eventify.controller;
+package eventify.pageControllers;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.ui.Model;
