+++ src/main/java/eventify/pageControllers/HomePageController.java
@@ -1,4 +1,4 @@
-package eventify.controller;
+package eventify.pageControllers;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
