+++ src/main/java/eventify/pageControllers/UserPageController.java
@@ -1,4 +1,4 @@
-package eventify.controller;
+package eventify.pageControllers;
 
 import eventify.dto.UserDTO;
 import eventify.model.User;
