+++ src/main/java/eventify/restControllers/EventController.java
@@ -1,4 +1,4 @@
-package eventify.controller;
+package eventify.restControllers;
 
 import eventify.dto.EventDTO;
 import eventify.mapper.Mapper;
