+++ src/main/java/eventify/restControllers/ReviewController.java
@@ -1,4 +1,4 @@
-package eventify.controller;
+package eventify.restControllers;
 
 import eventify.dto.ReviewDTO;
 import eventify.mapper.Mapper;
